fix(portals): tear down the assigned shortcut when a portal is deleted
Deleting a portal removed it from the store and search index but left
its global hotkey registered, so pressing the key kept firing the
deleted portal. Portal register/unregister logic now lives in
portalLifecycle.ts with a single deletePortal() that also unregisters
the item shortcut.

Part of #433
